fix: extract syncRemoteSkillsets so reconnect re-adds deleted platform skillsets
Moved the sync-remote logic from the inline route handler into
skillset-service.ts as a reusable `syncRemoteSkillsets` function.
The existing reconnect flow (use-platform-auth) already calls
POST /api/skillsets/sync-remote on successful auth — this ensures
platform skillsets that were manually deleted get re-added to
settings when the user reconnects.

Made-with: Cursor
